awid/pyproject.toml
[project]
name = "awid-service"
version = "0.1.2"
description = "awid.ai identity registry service and shared awid primitives package"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Juan Reyero", email = "juan@juanreyero.com" }
]
requires-python = ">=3.12"
dependencies = [
    "base58>=2.1.1",
    "dnspython>=2.8.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pgdbm>=0.4.1",
    "publicsuffix2>=2.20191221",
    "pynacl>=1.6.2",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "redis>=6.4.0",
    "typer>=0.16.1",
    "uvicorn[standard]>=0.35.0",
]

[project.urls]
Homepage = "https://github.com/awebai/aweb"
Repository = "https://github.com/awebai/aweb"
Documentation = "https://github.com/awebai/aweb/tree/main/docs"

[project.scripts]
awid = "awid_service.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/awid", "src/awid_service"]

[tool.hatch.build.targets.sdist]
exclude = [
    ".claude/",
    "AGENTS.md",
    "CLAUDE.md",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.2.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
